--- src/TransactionsContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from 'react'
-
-export const TransactionsContext = createContext<number[]>([])
