+++ src/components/Sumary/index.tsx
@@ -7,7 +7,7 @@ import { TransactionsContext } from "../../TransactionsContext";
 import { useContext } from "react";
 
 export function Summary() {
-  const data = useContext(TransactionsContext)
+  const transactions = useContext(TransactionsContext)
 
   return (
     <Container>
